refactored: fixing old components, removed name parameter that is not needed because vue does it automaticly
